requirements.txt
fastapi
uvicorn
pydantic
supabase-py
folium
phonenumbers
opencage
nltk
requests
regex
